requirements.txt
transformers 
pandas 
numpy 
matplotlib 
pillow